requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
tqdm>=4.62.0

# Vector databases
chromadb>=0.4.6
pinecone-client>=2.2.1

# Embedding and ML
scikit-learn>=1.0.0
sentence-transformers>=2.2.2
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
mistralai>=0.0.3
anthropic>=0.5.0
google-generativeai>=0.3.0

# Web scraping and PDF processing
selenium>=4.0.0
requests>=2.28.0
beautifulsoup4>=4.10.0
PyMuPDF>=1.20.0  # For PDF processing

# AWS
boto3>=1.24.0

# UI
streamlit>=1.22.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
tiktoken>=0.3.0  # For token counting with OpenAI

# Logging and monitoring
rich>=12.0.0

# Core FastAPI packages
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6  # For handling file uploads
pydantic>=2.4.2
starlette>=0.27.0

# For async operations
httpx>=0.24.1

# For testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
